requirements.txt
numpy
cmake
dlib
opencv-python
pytesseract
tensorflow-gpu==1.15.2
keras==2.3.1
Pillow
Pandas
Flask
mysql
mysql-connector
imutils
matplotlib
